Move node_hash_to_fixed_size shared function to a utils module
